requirements.txt
Django
gunicorn
django-cassandra-engine
cassandra-driver
cqlengine
blist
djangorestframework
requests
